+++ betty/cropper/utils/__init__.py
@@ -1,0 +1,6 @@
+import time
+
+
+# Necessary b/c Python<3.3 doesn't support datetime.timestamp()
+def seconds_since_epoch(when):
+    return int(time.mktime(when.timetuple()))