+++ betty/cropper/views.py
@@ -1,13 +1,17 @@
 import json
 from betty.conf.app import settings
 
-from django.http import Http404, HttpResponse, HttpResponseServerError, HttpResponseRedirect
+from django.http import (Http404, HttpResponse, HttpResponseNotModified,
+                         HttpResponseServerError, HttpResponseRedirect)
 from django.shortcuts import render
 from django.utils.cache import patch_cache_control
+from django.utils.http import http_date
+
 from django.views.decorators.cache import cache_control
 from six.moves import urllib
 
 from .models import Image, Ratio
+from .utils.http import check_not_modified
 from .utils.placeholder import placeholder
 
 logger = __import__('logging').getLogger(__name__)
@@ -76,6 +80,13 @@ def redirect_crop(request, id, ratio_slug, width, extension):
                                                        extension=extension))
 
 
+def _image_response(image_blob, extension):
+    resp = HttpResponse(image_blob)
+    resp["Content-Type"] = EXTENSION_MAP[extension]["mime_type"]
+    resp['Last-Modified'] = http_date()
+    return resp
+
+
 def crop(request, id, ratio_slug, width, extension):
     if ratio_slug != "original" and ratio_slug not in settings.BETTY_RATIOS:
         raise Http404
@@ -106,14 +117,18 @@ def crop(request, id, ratio_slug, width, extension):
         else:
             raise Http404
 
-    try:
-        image_blob = image.crop(ratio, width, extension)
-    except Exception:
-        logger.exception("Cropping error")
-        return HttpResponseServerError("Cropping error")
+    if check_not_modified(request=request, last_modified=image.last_modified):
+        # Avoid hitting storage backend on cache update
+        resp = HttpResponseNotModified()
+    else:
+        try:
+            image_blob = image.crop(ratio, width, extension)
+        except Exception:
+            logger.exception("Cropping error")
+            return HttpResponseServerError("Cropping error")
 
-    resp = HttpResponse(image_blob)
-    resp["Content-Type"] = EXTENSION_MAP[extension]["mime_type"]
+        resp = _image_response(image_blob, extension=extension)
+
     # Optionally specify alternate cache duration for non-breakpoint widths.
     # This is useful b/c cache flush callback only receives paths for known breakpoints, so this
     # allows non-standard widths to have a shorter cache time. This wouldn't be necessary if cache
@@ -128,7 +143,6 @@ def crop(request, id, ratio_slug, width, extension):
 
 # Legacy behavior -- originally these were just dropped on filesystem and let NGINX frontend serve
 # automatically via try-files.
-@cache_control(max_age=settings.BETTY_CACHE_CROP_SEC)
 def animated(request, id, extension):
 
     image_id = int(id.replace("/", ""))
@@ -141,12 +155,17 @@ def animated(request, id, extension):
     if not image.animated:
         raise Http404
 
-    try:
-        image_blob = image.get_animated(extension=extension)
-    except Exception:
-        logger.exception("Animated error")
-        return HttpResponseServerError("Animated error")
+    if check_not_modified(request=request, last_modified=image.last_modified):
+        # Avoid hitting storage backend on cache update
+        resp = HttpResponseNotModified()
+    else:
+        try:
+            image_blob = image.get_animated(extension=extension)
+        except Exception:
+            logger.exception("Animated error")
+            return HttpResponseServerError("Animated error")
 
-    resp = HttpResponse(image_blob)
-    resp["Content-Type"] = EXTENSION_MAP[extension]["mime_type"]
+        resp = _image_response(image_blob, extension=extension)
+
+    patch_cache_control(resp, max_age=settings.BETTY_CACHE_CROP_SEC)
     return resp