+++ tests/test_cropping.py
@@ -1,5 +1,8 @@
+import io
 import os
 
+from freezegun import freeze_time
+from PIL import Image as PILImage
 import pytest
 
 from django.core.files import File
@@ -11,6 +14,22 @@ from mock import patch
 
 
 TEST_DATA_PATH = os.path.join(os.path.dirname(__file__), 'images')
+
+
+@freeze_time('2016-05-02 01:02:03')
+@pytest.mark.django_db
+@pytest.mark.usefixtures("clean_image_root")
+def test_basic_cropping(settings, client, image):
+    settings.BETTY_CACHE_CROP_SEC = 600
+    res = client.get('/images/{}/1x1/200.jpg'.format(image.id))
+    assert res.status_code == 200
+    assert res['Cache-Control'] == 'max-age=600'
+    assert res['Last-Modified'] == "Mon, 02 May 2016 01:02:03 GMT"
+    assert res['Content-Type'] == "image/jpeg"
+
+    image_buffer = io.BytesIO(res.content)
+    img = PILImage.open(image_buffer)
+    assert img.size == (200, 200)
 
 
 @pytest.mark.django_db
@@ -273,3 +292,14 @@ def test_image_js_use_request_host(settings, client):
     res = client.get("/images/image.js", SERVER_NAME='alt.example.org')
     assert res.status_code == 200
     assert res.context['BETTY_IMAGE_URL'] == '//alt.example.org/images'
+
+
+@pytest.mark.django_db
+@pytest.mark.usefixtures("clean_image_root")
+def test_if_modified_since(settings, client, image):
+    settings.BETTY_CACHE_CROP_SEC = 600
+    res = client.get('/images/{}/1x1/300.jpg'.format(image.id),
+                     HTTP_IF_MODIFIED_SINCE="Sat, 01 May 2100 00:00:00 GMT")
+    assert res.status_code == 304
+    assert res['Cache-Control'] == 'max-age=600'
+    assert not res.content  # Empty content